Move product resolvers into dedicated modules
